Expense_TripDD/pyproject.toml
[project]
name = "expense-tripdd"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "docx-mailmerge>=0.5.0",
    "error-code>=0.0.11",
    "flask==3.1.3",
    "flask-sqlalchemy==3.1.1",
    "ipython-beautifulsoup>=0.3",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pymysql==1.1.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "requests>=2.32.5",
    "sqlalchemy==2.1.0b1",
    "success>=1.0.0",
    "yagmail>=0.15.293",
]
